Minor Tweaks, Did first letter for language
Created an draw_nodes boolean that decides wether to draw the node in the Editor with Nodes or in plain form.
When not pressing shift, only one node can be selected at a time
Fixed issue with debug window
When debug window can't send, it won't just call an Error. Instead, it prints out the message